typo tooltip
