--- src/Tooltip/index.d.ts
@@ -1,3 +1,0 @@
-import Tooltip, { TooltipProps } from './Tooltip';
-export default Tooltip;
-export { TooltipProps };
