--- src/Tooltip/index.js
@@ -1,2 +1,0 @@
-import Tooltip from './Tooltip';
-export default Tooltip;
