+++ src/Tooltip/index.ts
@@ -1,0 +1,3 @@
+import Tooltip from './Tooltip';
+export default Tooltip;
+export { TooltipProps } from './Tooltip';
